Modified fire-cat, gige-cat to use consistent args
In most comma/snark programs, command line arguments are separated by
semicolons. e.g.

    cv-cat --input="rows=500;cols=1000;no-header;type=3ub"

But fire-cat and gige-cat required parameters in the form

    fire-cat --config="output-type=RGB,guid=12345"

To make these apps consistent with the rest of comma and snark, the "--config"
and "--set" arguments of fire-cat and gige-cat respectively have been changed
to expect semicolon-separated name-value pairs.