graphics: ccmake will use default Qt version (rather than Qt5)
Previously, ccmake would try to find if Qt5 is installed and then if found it would default to
use Qt5. It would also set up the environment to use Qt5, so that even if the
user selects Qt4 later, some applications (moc, in particular) would still use Qt5
version, which leads to incompatibility and compile errors. In the new version,
ccmake will try to determine which version of Qt is enabled and then use this version.