actuators/universal_robots: disabled tests that fail on the test server
Building universal_robots was enabled by default in the previous commit in order to allow the tests to run on the test server.
However, the test still failed because the required application (ur-arm-command) was not installed.
Apparently, the test server reuses the previous default setting where building universal_robots is disabled.
The proper fix would be to modify the scripts on the test server such that ccmake picks up the new default settings correctly.