actuators/universal_robots: went back to using io-publish in the test
version of nc installed on the test server does not work as intended (there is no socat there either), so the test uses io-publish.
ur-arm test is disabled until the necessary changes on the test server that make io-publish available there for snark.
pkill is used in the test to kill all created background processes cleanly ( while, sleep, io-publish).
option -4 is added to nc calls in ur-arm get and send functions for compatibility with the nc version installed on the test server.