sensors/dc1394/applications/fire-cat.cpp: added capability to read config from json, xml, and path-value
The format of the config file will be guessed based on the input; json, xml, and path-value are accepted.
Simplified ini format can also be used, e.g. no sections or comments ( and no '/' symbols in keys), and will be treated the same as path-value format.
Running fire-cat will no longer drop a fire-cat.ini file in the current directory (see verbose help, for default values of camera parameters).