actuators/universal_robots/applications/ur-arm: auto-init checks if the arm is in home position
Use option --force to force auto-init if the arm is not in home position.
Renamed release-brakes -> unlock-breaks.
Output from all operations goes to stderr, except for the config produced by init, which is output to stdout.
Operations such as on, init, auto-init, and stow yield 0 exit code if they succeed.