control/applications/control-follow-waypoints: restored daemonize for feedback
Trapping signals in the wrapper srcipt lead to some signals not
propagaring to processes run by comma_execute_and_wait.
A heavy solution is to use comma_execute_and_wait in the wrapper script.
A light solution (preferred in this commit) is to get rid of signal
trapping in the wrapper script.