math/spherical_geometry/coordinates: refactor for review and usability:  - give symbolic names to constants  - elaborate in comments all the estimates and approximations that went    into the lower estimate for the great circle arc length vs delta in    longitude  - make 3 decision branches explicit
Also, add debugging code to count the number of calls into each of the 3
decision branches to (i) measure comparative performance of each branch, and
(ii) to see how often they are invoked in real code. The answers are:
 - shortcuts are about 50 times faster than calculating angle() directly
 - the difference between the latitude and longitude shortcuts is under 25%
 - the alternative longitude shortcut using trigonometry (commented out as
   of this commit) is over 10 times slower than more crude estimate used in
   the ordinary shortcut (still 4 times faster than angle calculations)
 - the exact branch is invoked in about 0.7% of all cases in a sample run of
   a larger flight planning model using the coordinates code

Therefore, shortcuts are vital and give a major speed up, and trigonometry
shall not be used in the longitude shortcut.