points-to-ros: downgrade warning about conversion of type t to verbose only
It's a bit misleading, because when t is copied to the header (where it
matters) it's correctly copied in a ROS compatible way.

We don't really need to copy t and block in the data because they are in the
head but that's a bigger change and not for today.