nmea-to-csv: add handling of date from RMC messages.
GGA messages give only "time-of-day", so when running nmea-to-csv
on logs with no ZDA mesages, nmea-to-csv would use the local system
time.

This is probably never what you want, so instead nmea-to-csv now
retrieves date information from RMC messages. Note that csv output
is now supressed whilst waiting for the first RMC message.