darknet-cat: refactored to fit alexeyab/darknet.
https://github.com/alexeyab/darknet is now the canonical version of
the darknet library. It has support for newer models and GPUs.

The ABI changes frequently, which means it is tricky to keep manual
python bindings in sync -- instead we now require that the darknet
libraries are installed to a global location and then darknet-cat
accesses the python api, and compiled libdarknet.so from there.

This will require a recent version of alexeyab/darknet (> 8 Aug 2020,
commit 9ecc4f7d6b24dec4d56a1d5c480fe842625dc57b).

Another significant improvement is the python code now redirects
darknet's stdout to stderr, so that darknet-cat can be pipe-compatible.
Previously this required changes to darknet itself.