ros-image: only unblock SIGPIPE when reading from a bag file
not when using ros::spin() or ros::spinOnce().

Certainly for ros::spin(), when publishing a high-bandwidth source we see
_lots_ of SIGPIPE's. ROS seems not to know how to handle them so just blocks
the signal. See signal.txt for more detail.

This might mean certain types of use, publishing through a pipe, won't work.
Or at the very least will need to be revisited.

This reverts commit 47f60cea (and removes the failing tests that led to that
commit)